# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Flowra API
  description: |
    Chat System with Task Management API.

    This API provides comprehensive functionality for:
    - User authentication (Keycloak SSO integration)
    - Workspace management with role-based access control
    - Real-time chat with participants management
    - Task management with kanban-style workflows
    - Message handling with attachments and reactions
    - Notification system
    - WebSocket real-time communication

    ## Authentication

    All endpoints (except `/auth/login` and `/auth/register`) require JWT authentication.
    Include the JWT token in the `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    ## Rate Limiting

    - Default: 100 requests per minute per user
    - Auth endpoints: 10 requests per minute per IP
    - WebSocket: 60 messages per minute per connection

    ## Error Handling

    All errors follow a consistent format with `code` and `message` fields.
  version: 1.0.0
  contact:
    name: Flowra Development Team
    email: dev@flowra.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.flowra.com/api/v1
    description: Production server
security:
  - bearerAuth: []
...
paths:
...
  /workspaces/{workspace_id}:
    get:
      tags:
        - Workspaces
      summary: Get workspace details
      description: Returns detailed information about a specific workspace
      operationId: getWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdPath"
      responses:
        "200":
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    put:
      tags:
        - Workspaces
      summary: Update workspace
      description: Updates workspace settings. Requires admin or owner role.
      operationId: updateWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkspaceRequest"
            example:
              name: "Updated Workspace Name"
              description: "Updated description"
      responses:
        "200":
          description: Workspace updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"

    delete:
      tags:
        - Workspaces
      summary: Delete workspace
      description: Permanently deletes a workspace. Requires owner role.
      operationId: deleteWorkspace
      parameters:
        - $ref: "#/components/parameters/WorkspaceIdPath"
      responses:
        "204":
          description: Workspace deleted successfully
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "403":
          $ref: "#/components/responses/ForbiddenError"
        "404":
          $ref: "#/components/responses/NotFoundError"
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login

  # ============================================
  # Parameters
  # ============================================
  parameters:
    WorkspaceIdPath:
      name: workspace_id
      in: path
      required: true
      description: Workspace ID
      schema:
        type: string
        format: uuid
...
  responses:
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: "UNAUTHORIZED"
              message: "Authentication required"
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: "FORBIDDEN"
              message: "You don't have permission to perform this action"
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: "NOT_FOUND"
              message: "Resource not found"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            success: false
            error:
              code: "VALIDATION_ERROR"
              message: "Invalid request data"
...
  schemas:
...
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Error code
              example: "VALIDATION_ERROR"
            message:
              type: string
              description: Human-readable error message
              example: "Invalid request data"
            details:
              type: object
              description: Additional error details
              additionalProperties:
                type: string

    # Authentication schemas
...
    UpdateWorkspaceRequest:
      type: object
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
    WorkspaceResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
          properties:
            id:
              type: string
              format: uuid
            name:
              type: string
            description:
              type: string
            owner_id:
              type: string
              format: uuid
            member_count:
              type: integer
            created_at:
              type: string
              format: date-time
            updated_at:
              type: string
              format: date-time
...